feat(ehr): removing method + updating utils
Ref: #1040
